updates for app rui support files

+++ rbx_drivers/rbx_ardupilot_discovery.py
@@ -11,13 +11,18 @@
 # NEPI Auto Discovery Script for Sealite devices
 
 import os
+#NEPI_BASE_NAMESPACE = '/nepi/s2x/'
+#os.environ["ROS_NAMESPACE"] = NEPI_BASE_NAMESPACE[0:-1]
+import rospy
 import subprocess
 import time
-import rospy
 import serial
 import serial.tools.list_ports
+import socket
 
 from nepi_ros_interfaces.srv import LSXCapabilitiesQuery
+from mavros_msgs.srv import VehicleInfoGet
+from mavros_msgs.msg import VehicleInfo
 
 from nepi_edge_sdk_base import nepi_nex
 
@@ -40,8 +45,8 @@ TEST_NEX_DICT = {
     'driver_class_name':  "None",
     'driver_interfaces': ['USBSERIAL'],
     'driver_options_1': ['Real','SITL'],
-    'driver_default_option_1': 'Real',
-    'driver_set_option_1': 'Real',
+    'driver_default_option_1': 'SITL',
+    'driver_set_option_1': 'SITL',
     'driver_options_2': [],
     'driver_default_option_2': 'None',
     'driver_set_option_2': 'None',
@@ -69,7 +74,8 @@ class ArdupilotDiscovery:
   active_devices_dict = dict()
   node_name = "ardupilot"
   baudrate_list = [57600]
-  sitl_addr_list = ['127.0.0.0']
+  ip_addr_list = ['127.0.0.1']
+  ip_port_list = ['5760']
   ################################################          
   def __init__(self):
     self.log_name = PKG_NAME.lower() + "_discovery" 
@@ -120,20 +126,17 @@ class ArdupilotDiscovery:
             if success:
               self.active_paths_list.append(path_str)
     elif self.system_type == 'SITL':
-        self.path_list = self.sitl_addr_list
-        for path_str in self.path_list:
-          valid_path = True
-          for id in self.ignore_id_list:
-            if path_str.find(id) != -1 or path_str in self.active_paths_list:
-              valid_path = False
-          if valid_path:
-            #rospy.logwarn(self.log_name + ": Looking for path: " + path_str)
-            #rospy.logwarn(self.log_name + ": In path_list: " + str(self.active_paths_list))
-            found = self.checkForSITLDevice(path_str)
-            if found:
-              success = self.launchSITLDeviceNode(path_str)
-              if success:
-                self.active_paths_list.append(path_str)
+        for ip_addr_str in self.ip_addr_list:
+          for ip_port_str in self.ip_port_list:
+            path_str = ip_addr_str + "_" + ip_port_str
+            if path_str not in self.active_paths_list:
+              #rospy.logwarn(self.log_name + ": Looking for path: " + path_str)
+              #rospy.logwarn(self.log_name + ": In path_list: " + str(self.active_paths_list))
+              found = self.checkForSITLDevice(path_str)
+              if found:
+                success = self.launchSITLDeviceNode(path_str)
+                if success:
+                  self.active_paths_list.append(path_str)
 
     else:
       return self.active_paths_list
@@ -223,9 +226,7 @@ class ArdupilotDiscovery:
     return found_device
 
 
-  def checkForSITLDevice(self,path_str):
-    found_device = False
-    return found_device
+  
 
 
   def checkOnDevice(self,path_str):
@@ -282,26 +283,27 @@ class ArdupilotDiscovery:
           except:
             pass        
             
+          if ardu_subproc is not None:
+            if ardu_subproc.poll() is None:
+              rospy.logwarn(self.log_name + ": " + "Issuing sigterm to process for %s as part of node purging", "ardupilot_node")
+              ardu_subproc.kill()
+              # Turns out that is not always enough to get the node out of the ros system, so we use rosnode cleanup, too
+              # rosnode cleanup won't find the disconnected node until the process is fully terminated
+              try:
+                ardu_subproc.wait(timeout=10)
+              except:
+                pass
 
-          if ardu_subproc.poll() is None:
-            rospy.logwarn(self.log_name + ": " + "Issuing sigterm to process for %s as part of node purging", "ardupilot_node")
-            ardu_subproc.kill()
-            # Turns out that is not always enough to get the node out of the ros system, so we use rosnode cleanup, too
-            # rosnode cleanup won't find the disconnected node until the process is fully terminated
-            try:
-              ardu_subproc.wait(timeout=10)
-            except:
-              pass
-
-          if fgps_subproc.poll() is None:
-            rospy.logwarn(self.log_name + ": " + "Issuing sigterm to process for %s as part of node purging", "fgps_gps_node")
-            fgps_subproc.kill()
-            # Turns out that is not always enough to get the node out of the ros system, so we use rosnode cleanup, too
-            # rosnode cleanup won't find the disconnected node until the process is fully terminated
-            try:
-              fgps_subproc.wait(timeout=10)
-            except:
-              pass
+          if fgps_subproc is not None:
+            if fgps_subproc.poll() is None:
+              rospy.logwarn(self.log_name + ": " + "Issuing sigterm to process for %s as part of node purging", "fgps_gps_node")
+              fgps_subproc.kill()
+              # Turns out that is not always enough to get the node out of the ros system, so we use rosnode cleanup, too
+              # rosnode cleanup won't find the disconnected node until the process is fully terminated
+              try:
+                fgps_subproc.wait(timeout=10)
+              except:
+                pass
         
           cleanup_proc = subprocess.Popen(['rosnode', 'cleanup'], stdin=subprocess.PIPE)
           try:
@@ -315,8 +317,8 @@ class ArdupilotDiscovery:
   def launchRealDeviceNode(self, path_str):
     success = False
     if self.comp_id != None and self.sys_id_mavlink != None and self.addr_str != None:
-          path_str_short = path_str.split('/')[-1]
-          mavlink_node_name = "mavlink_" + path_str_short
+          ip_str_short = path_str.split('/')[-1]
+          mavlink_node_name = "mavlink_" + ip_str_short
           mavlink_node_namespace = self.base_namespace + mavlink_node_name
           rospy.loginfo("Mavlink_AD: Starting mavlink node setup: " + mavlink_node_name)
           # Load the proper configs for APM
@@ -340,14 +342,14 @@ class ArdupilotDiscovery:
           mav_subproc = subprocess.Popen(node_run_cmd)
          
           # Start the ardupilot RBX node for this mavlink connection
-          ardu_node_name = "ardupilot_" + path_str_short
+          ardu_node_name = "ardupilot_" + ip_str_short
           rospy.loginfo("Mavlink_AD: " + "Starting ardupilot rbx node: " + ardu_node_name)
           processor_run_cmd = ["rosrun", "nepi_drivers", "rbx_ardupilot_node.py",
                                 "__name:=" + ardu_node_name, f"__ns:={self.base_namespace}"]
           ardu_subproc = subprocess.Popen(processor_run_cmd)
 
           #Start the an RBX fake gps for this ardupilot node
-          fgps_node_name = "fake_gps_" + path_str_short
+          fgps_node_name = "fake_gps_" + ip_str_short
           rospy.loginfo("Mavlink_AD: " + "Starting fake gps rbx node: " + fgps_node_name)
           processor_run_cmd = ["rosrun", "nepi_edge_sdk_base", "fake_gps.py",
                                 "__name:=" + fgps_node_name, f"__ns:={self.base_namespace}"]
@@ -360,20 +362,143 @@ class ArdupilotDiscovery:
             rospy.wait_for_service(vehicle_info_service_name, timeout=10) # TODO: 10 seconds always sufficient for the driver?
             # No exception, all good
             device_entry = dict()
-            device_entry["sysid"] = sys_id_mavlink
-            device_entry["compid"] =  comp_id   
+            device_entry["sysid"] = self.sys_id_mavlink
+            device_entry["compid"] =  self.comp_id   
             device_entry["node_name"] =  mavlink_node_name
             device_entry["mavlink_subproc"] = mav_subproc
             device_entry["ardu_subproc"] = ardu_subproc
             device_entry["fgps_subproc"] = fgps_subproc
 
-            active_devices_dict[path_str] = device_entry
+            self.active_devices_dict[path_str] = device_entry
             success = True
+          except Exception as e:
+            rospy.logerr(self.log_name + ": Failed to start " + mavlink_node_name + " " + str(e))
+    return success
+
+
+
+  def checkForSITLDevice(self,path_str):
+    found_device = False
+    [ip_addr_str,ip_port_str] = path_str.split("_")
+    ip_addr_str_list = ip_addr_str.split('.')
+    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+    result = sock.connect_ex((ip_addr_str,int(ip_port_str)))
+    if result == 0:
+        found_device = True
+        rospy.logwarn("Mavlink_AD: Found device on ip address: " + ip_addr_str + " port: " + ip_port_str + " is open")
+        sock.close()
+    else:
+      rospy.logwarn("Mavlink_AD: Did not find device on ip address: " + ip_addr_str + " port: " + ip_port_str)
+    return found_device
+    
+
+
+
+  def launchSITLDeviceNode(self, path_str):
+    success = False
+    [ip_addr_str,ip_port_str] = path_str.split("_")
+    ip_addr_str_list = ip_addr_str.split('.')
+    ip_str_short = ''.join(ip_addr_str_list)
+
+
+    rospy.logdebug("Mavlink_AD: : Connecting to ardupilot simulator at ip address: " + ip_addr_str + " port: " + ip_port_str)
+
+    mavlink_node_name = "mavlink_"  + ip_str_short
+    mavlink_node_namespace = self.base_namespace + mavlink_node_name
+    rospy.loginfo("Mavlink_AD: Starting mavlink node setup: " + mavlink_node_name)
+    # Load the proper configs for APM
+    rosparam_load_cmd = ['rosparam', 'load', '/opt/nepi/ros/share/mavros/launch/apm_pluginlists.yaml', mavlink_node_namespace]
+    subprocess.run(rosparam_load_cmd)
+    rosparam_load_cmd = ['rosparam', 'load', '/opt/nepi/ros/share/mavros/launch/apm_config.yaml', mavlink_node_namespace]
+    subprocess.run(rosparam_load_cmd)
+    # Adjust the timesync_rate to cut down on log noise
+    rospy.set_param(mavlink_node_namespace + '/conn/timesync_rate', 1.0)
+    # Allow the HIL plugin. Disabled in apm configs for some reason
+    
+    # Launch Mavlink Node
+    rospy.loginfo("Mavlink_AD: Launching mavlink node: " + mavlink_node_name)
+    #fcu_url = "udp://14555@014550"
+    #fcu_url = "udp://14555@0.0.0.0.0:14550"
+    #fcu_url = "udp://" + ip_addr_str + ":14555@0.0.0.0.0:14550"
+    #fcu_url = "tcp://" + ip_addr_str + ":5760"
+    #fcu_url = "udp://:14540@14557"
+    fcu_url = "udp://:14540@127.0.0.1:14557"
+    #fcu_url = "tcp://" + ip_addr_str + ":" + ip_port_str
+    #fcu_url = "udp://" + ip_addr_str + ":" + ip_port_str
+    gcs_url = ""
+    node_run_cmd = ['rosrun', 'mavros', 'mavros_node', '__name:=' + mavlink_node_name, '_fcu_url:=' + fcu_url ] #, '_gcs_url:=' + gcs_url] 
+    mav_subproc = subprocess.Popen(node_run_cmd)
+    time.sleep(5)
+    vehicle_info_service_name = mavlink_node_namespace + '/vehicle_info_get'
+    try:
+      rospy.wait_for_service(vehicle_info_service_name, timeout=10) # TODO: 10 seconds always sufficient for the driver?
+    except Exception as e:
+      rospy.logerr(self.log_name + ": Failed to start " + mavlink_node_name + " " + str(e))
+
+    # Now check that the node is actually responsive
+    # Use a service call so that we can provide are assured of synchronous response
+    vehicle_info_service_name = full_node_name + '/vehicle_info_get'
+    vehicle_info_query = rospy.ServiceProxy(vehicle_info_service_name, VehicleInfoGet)
+    try:
+      # We don't actually care about the contents of the response at this point, but we might in the future for
+      # additional aliveness check logic:
+      #response = capability_service()
+      response = vehicle_info_query(sysid=sysid, compid=compid, get_all=True)
+      vehicle_list = response.vehicles
+      if len(vehicle_list):
+        vehicle = vehicle_list[0]
+        rospy.loginfo("Mavlink_AD: " + "Found robot info: " + str(vehicle))
+        sysid = vehicle.sysid
+        compid = vehicle.conpid
+      success = True
+    except Exception as e: # Any exception indicates that the service call failed
+      rospy.logwarn(self.log_name + ": " + " Node %s is no longer responding to vehicle info queries (%s)", ros_node_name, str(e))
+      purge_node = True  
+
+      # No exception, all good
+      device_entry = dict()
+      device_entry["sysid"] = 1
+      device_entry["compid"] =  1   
+      device_entry["node_name"] =  mavlink_node_name
+      device_entry["mavlink_subproc"] = mav_subproc
+      device_entry["ardu_subproc"] = None
+      device_entry["fgps_subproc"] = None
+
+
+    if success:
+      # Start the ardupilot RBX node for this mavlink connection
+      ardu_node_name = "ardupilot_" + ip_str_short
+      rospy.loginfo("Mavlink_AD: " + "Starting ardupilot rbx node: " + ardu_node_name)
+      processor_run_cmd = ["rosrun", "nepi_drivers", "rbx_ardupilot_node.py",
+                            "__name:=" + ardu_node_name, f"__ns:={self.base_namespace}"]
+      ardu_subproc = subprocess.Popen(processor_run_cmd)
+      device_entry["ardu_subproc"] = ardu_subproc
+      self.active_devices_dict[path_str] = device_entry
+      # And make sure it actually starts up fully by waiting for a guaranteed service
+    else:
+        if mavlink_subproc.poll() is None:
+          rospy.logwarn(self.log_name + ": " + "Issuing sigterm to process for %s as part of node purging", ros_node_name)
+          mavlink_subproc.kill()
+          # Turns out that is not always enough to get the node out of the ros system, so we use rosnode cleanup, too
+          # rosnode cleanup won't find the disconnected node until the process is fully terminated
+          try:
+            mavlink_subproc.wait(timeout=10)
           except:
-            rospy.logerr(self.log_name + ": Failed to start " + mavlink_node_name)
+            pass              
     return success
 
 
   def launchSTILDeviceNode(self, path_str):
     success = False
     return success
+
+
+#########################################
+# Main
+#########################################
+if __name__ == '__main__':
+  from nepi_edge_sdk_base import nepi_ros
+  bn = nepi_ros.get_base_namespace()
+  ardu = ArdupilotDiscovery()
+  print(ardu.discoveryFunction([],[],bn))
+
